Check for updated files in the backend in the retrieve* functions.
This is done by doing a lightweight HEAD request and then checking
whether the modification time occurs after the age of the cached file;
if so, we overwrite the cache with the new remote content.

We make sure to set the modification time correctly on all downloaded
files to ensure that we can accurately determine their age in the cache.